+++ libavcodec/vc1dsp.h
@@ -31,7 +31,7 @@
 #include "dsputil.h"
 #include "h264chroma.h"
 
-typedef void (*vc1op_pixels_func)(uint8_t *block/*align width (8 or 16)*/, const uint8_t *pixels/*align 1*/, int line_size, int h);
+typedef void (*vc1op_pixels_func)(uint8_t *block/*align width (8 or 16)*/, const uint8_t *pixels/*align 1*/, ptrdiff_t line_size, int h);
 
 typedef struct VC1DSPContext {
     /* vc1 functions */
